+++ React-Catatan-Belanja/src/components/Header.jsx
@@ -1,0 +1,5 @@
+export default function Header() {
+    return (
+      <h1>Catatan Belanjaku 📝</h1>
+    );
+  }
